App/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_api_key==3.0.0
Requests==2.32.3
scapy==2.6.1
gunicorn==23.0.0
psycopg2==2.9.10
greenlet==3.1.1
gevent==24.11.1